+++ src/app/view-config-utils.test.fixtures.js
@@ -390,6 +390,21 @@ export const initializedViewConfig = {
     spatialRotation: {
       A: 0,
     },
+    spatialRotationOrbit: {
+      A: 0,
+    },
+    spatialOrbitAxis: {
+      A: 'Y',
+    },
+    spatialRotationX: {
+      A: null,
+    },
+    spatialRotationY: {
+      A: null,
+    },
+    spatialRotationZ: {
+      A: null,
+    },
     spatialTargetX: {
       A: 3800,
     },
@@ -397,10 +412,13 @@ export const initializedViewConfig = {
       A: -900,
     },
     spatialTargetZ: {
-      A: 0,
+      A: null,
     },
     spatialZoom: {
       A: -4.4,
+    },
+    spatialAxisFixed: {
+      A: false,
     },
     additionalCellSets: {
       A: null,
@@ -506,9 +524,15 @@ export const initializedViewConfig = {
         spatialMoleculesLayer: 'A',
         spatialNeighborhoodsLayer: 'A',
         spatialRotation: 'A',
+        spatialRotationOrbit: 'A',
+        spatialOrbitAxis: 'A',
+        spatialRotationX: 'A',
+        spatialRotationY: 'A',
+        spatialRotationZ: 'A',
         spatialTargetX: 'A',
         spatialTargetY: 'A',
         spatialTargetZ: 'A',
+        spatialAxisFixed: 'A',
         spatialZoom: 'A',
         additionalCellSets: 'A',
         moleculeHighlight: 'A',