+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 - Use component height for determining 3D sizing.
   - Use state rather than computing every render.
+- No selection of cells in bitmask now results in default "grey" color instead of the last selected cell set.
 
 ## [1.1.11](https://www.npmjs.com/package/vitessce/v/1.1.11) - 2021-06-25