+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 ### Added
 - Added a new documentation site.
+
+### Changed
+- Use component height for determining 3D sizing.
+  - Use state rather than computing every render.
+
+## [1.1.11](https://www.npmjs.com/package/vitessce/v/1.1.11) - 2021-06-25
+
+### Added
+- Global visibility button next to name in layer controller.
+- Volumetric ray casting from `Viv`
+    - Upgrade `Viv` to 0.10.4
+    - Add new coordination types `spatialRotationX` `spatialRotationY` `spatialRotationZ` `spatialRotationOrbit` `spatialOrbitAxis` and update `spatialRasterLayers` with new parts
+    - Add spatial view state coordination types to `LayerController`
+    - Update UI for `LayerController`
+- Global visbiility prop per layer in `spatialRasterLayers`.
+- Add indication to Y axis title of cell set expression violin plot when log-transformation is active.
+
+### Changed
+- Cache cell set polygon outputs and do not calculate them unless requested.
+  - Modify the cache to use an array of tuples, since using an array as an object key results in conversion to string.
+- Clean up `getFlatArrDecompressed` fetching.
+- Fix bitmask remove button style.
+- Don't show 3D dropdown if only 2D is available.
+- Don't show Volume tab (or any tabs) when 3D is not available.
+
+## [1.1.10](https://www.npmjs.com/package/vitessce/v/1.1.10) - 2021-05-19
+
+### Added
 - Add support for bitmasks to `Spatial` component and raster schema.
 - Worker pool for processing heatmap tiles.
 
@@ -15,6 +43,8 @@
 - Fix bug where polygons or centroids would show under the `bitmask`.
 - `bitmask` color texture creation assumed that `cellColors` prop was only rgb, but it can be rgba.
 - Fix bug where quadtree wouldn't work with only scatterplot.
+- Fix controller padding bug.
+- Ensure `VlenUtf8` filter is only used/checked when necessary.
 
 ## [1.1.9](https://www.npmjs.com/package/vitessce/v/1.1.9) - 2021-05-07